Break writegpt() apart
This is a precursor to functionality to be added to mergegpt.c.  Now
there is _writegpt() for the low-level writing the data to media and
gpt_entries2raw() for the one preparation stage.  This is meant as
lower-level for entries which have already had their CRC computed (or
rigged so it computes to a certain value).