Check-in of small steps towards uncompressing data for testing
These changes are certain to be needed in the near future for working
with zlib as experimentation has shown.  Since they keep being pulled out
then reintroduced due to other work on this project, get these bits in to
Git now, rather than later.

Add a large note to kdz.c about which portions of KDZ files to check and
which are likely to need to be preserved.