Implement test_kdzfile(), aiming to test whether a KDZ applies to the device
There is great need to be able to test whether a given KDZ is likely to
successfully apply to a given device.  Applying a bad KDZ is likely to
cause a device not to boot ("brick").  The relationship between the KDZ
and on-device structures is quite complicated.  This is looks plausible,
I just hope my guesses were reasonable.