Add support for runtime linking of libraries
The Android NDK provides access to a limited selection of Android
libraries.  Additional libraries must be linked in using special
procedures instead of -l<lib>.  The approach used was to access the
runtime dynamic linker (dlopen()/dlsym()/dlclose()).  This is an
appropriate solution, but for building as part of recovery (TWRP/other)
linking during the build process is better.

Add support for disabling use of dlopen()/dlsym()/dlclose().  Compile
with BUILDTIME_LINK_LIBS and then add -lcrypto -lselinux during link.